Design Redeem Page
A frontend implementation of how the redeem page works has been implemented. The values on the page refreshes everytime you go back to it right now, but gives an example of how a template redeem function can work once we implement it from the backend, and we can rework the current UI components a bit as required.