Design Add Emissions Page
Frontend functionality of how the user can update their emissions manually, how the graph is to update and how the total emissions is to be reflected is there. Went for a bottom sheet modal implementation for the user to input their emissions. The overall UI and the functionality of how this page works can be tweaked a little.